Enable compiling GN by default.
This removes the call to RestoreDefaultExceptionHandler on Mac. This isn't exported by base so breaks on the component build. IT doesn't matter since this is only related to breakpad which we don't use, so I just removed the call.

BUG=288991
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/25153002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@226050 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9aa0bb10f4f0c035609435efb6762f673d30476b

+++ tools/gn/err.cc
@@ -41,7 +41,7 @@ void FillRangeOnLine(const LocationRange& range, int line_number,
 
   int end_char;
   if (range.end().line_number() > line_number)
-    end_char = line->size();  // Ending is non-inclusive.
+    end_char = static_cast<int>(line->size());  // Ending is non-inclusive.
   else
     end_char = range.end().char_offset() - 1;
