+++ tools/gn/standard_out.cc
@@ -74,7 +74,8 @@ void OutputString(const std::string& output, TextDecoration dec) {
   }
 
   DWORD written = 0;
-  ::WriteFile(hstdout, output.c_str(), output.size(), &written, NULL);
+  ::WriteFile(hstdout, output.c_str(), static_cast<DWORD>(output.size()),
+              &written, NULL);
 
   if (is_console)
     ::SetConsoleTextAttribute(hstdout, default_attributes);
