+++ tools/gn/trace.cc
@@ -271,5 +271,6 @@ void SaveTraces(const base::FilePath& file_name) {
   out << "]}";
 
   std::string out_str = out.str();
-  file_util::WriteFile(file_name, out_str.data(), out_str.size());
+  file_util::WriteFile(file_name, out_str.data(),
+                       static_cast<int>(out_str.size()));
 }
